+++ src/test/java/com/nextstep/racingcar/domain/ParserTest.java
@@ -1,17 +1,14 @@
-package com.nextstep.racingcar;
+package com.nextstep.racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.List;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class ParserTest {
     @DisplayName("이름 , 단위로 분리 기능")
